third_party/WebKit/LayoutTests/accessibility/focus-action-clicks-non-focusable-element.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div role="grid" aria-activedescendant="" tabindex="-1" id="container-with-active-descendant">
  <div role="row">
    <div role="gridcell id="cell-with-active-descendant">Cell</div>
    <div role="gridcell id="clickable-cell-with-active-descendant">Cell</div>
  </div>
</div>
<div id="non-focusable-element">Div</div>
<button id="focusable-element">Button</button>

<script>
test(function() {
  if (document.activeElement)
    document.activeElement.blur();

    var container = document.getElementById('container-with-active-descendant');
  var accessibilityFocusable = document.getElementById('cell-with-active-descendant');
  var axFocusable = accessibilityController.accessibleElementById('cell-with-active-descendant');

  var gotEvent = false;
  container.addEventListener('click', function(e) {
    gotEvent = true;
  });
  axFocusable.takeFocus();

  assert_true(gotClickEvent);
  assert_not_equals(document.activeElement, accessibilityFocusable);
  accessibilityFocusable.display = 'none';
}, 'A click event should be dispatched if a non-focusable element can take accessibility focus via aria-activedescendant.');

test(function() {
  if (document.activeElement)
    document.activeElement.blur();

  var accessibilityFocusable = document.getElementById('clickable-cell-with-active-descendant');
  var axFocusable = accessibilityController.accessibleElementById('cell-with-active-descendant');

  var gotEvent = false;
  accessibilityFocusable.addEventListener('click', function(e) {
    gotEvent = true;
  });
  axFocusable.takeFocus();

  assert_false(gotClickEvent);
  assert_not_equals(document.activeElement, accessibilityFocusable);
  accessibilityFocusable.display = 'none';
}, 'A click event should not be dispatched if a click handler is attached directly to an element that can take accessibility focus via aria-activedescendant.');

test(function() {
  if (document.activeElement)
    document.activeElement.blur();

  var nonFocusable = document.getElementById('non-focusable-element');
  var axNonFocusable = accessibilityController.accessibleElementById('non-focusable-element');

  var gotEvent = false;
  nonFocusable.addEventListener('click', function(e) {
    gotEvent = true;
  });
  axNonFocusable.takeFocus();

  assert_false(gotClickEvent);
  assert_not_equals(document.activeElement, nonFocusable);
  nonFocusable.display = 'none';
}, 'A click event should not be dispatched if an element cannot take accessibility focus.');

test(function() {
  if (document.activeElement)
    document.activeElement.blur();

  var focusable = document.getElementById('focusable-element');
  var axFocusable = accessibilityController.accessibleElementById('focusable-element');

  var gotEvent = false;
  focusable.addEventListener('click', function(e) {
    gotEvent = true;
  });
  axFocusable.takeFocus();

  assert_false(gotClickEvent);
  assert_equals(document.activeElement, focusable);
  focusable.display = 'none';
}, 'A click event should not be dispatched if an element is focusable.');
</script>
